test: continue coordinator test decomposition
